+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/coding/magicnumber/InputIntMethodAnnotation.java
@@ -1,4 +1,4 @@
-package com.puppycrawl.tools.checkstyle.checks.coding;
+package com.puppycrawl.tools.checkstyle.checks.coding.magicnumber;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
